+++ db/query_test.go
@@ -1,166 +1,360 @@
 package db
 
 import (
-	"errors"
-	"reflect"
 	"sort"
 	"strings"
 	"testing"
-
-	core "github.com/textileio/go-threads/core/db"
-	"github.com/textileio/go-threads/util"
 )
 
-type book struct {
-	ID     core.InstanceID
-	Title  string
-	Author string
-	Meta   bookStats
-}
+const (
+	testQuerySchema = `{
+		"$schema": "http://json-schema.org/draft-04/schema#",
+		"$ref": "#/definitions/book",
+		"definitions": {
+		   "book": {
+			  "required": [
+				 "ID",
+				 "Title",
+				 "Author",
+				 "Meta"
+			  ],
+			  "properties": {
+				 "Author": {
+					"type": "string"
+				 },
+				 "Banned": {
+					"type": "boolean"
+				 },
+				 "ID": {
+					"type": "string"
+				 },
+				 "Meta": {
+					"$schema": "http://json-schema.org/draft-04/schema#",
+					"$ref": "#/definitions/bookStats"
+				 },
+				 "Title": {
+					"type": "string"
+				 }
+			  },
+			  "additionalProperties": false,
+			  "type": "object"
+		   },
+		   "bookStats": {
+			  "required": [
+				 "TotalReads",
+				 "Rating"
+			  ],
+			  "properties": {
+				 "Rating": {
+					"type": "number"
+				 },
+				 "TotalReads": {
+					"type": "integer"
+				 }
+			  },
+			  "additionalProperties": false,
+			  "type": "object"
+		   }
+		}
+	 }`
+)
 
-type bookStats struct {
-	TotalReads int
-	Rating     float64
-}
-
-type queryTest struct {
+type jsonQueryTest struct {
 	name    string
 	query   *Query
-	resIdx  []int // expected idx results from sample data
+	resIdx  []int
 	ordered bool
 }
 
 var (
-	sampleData = []book{
-		{Title: "Title1", Author: "Author1", Meta: bookStats{TotalReads: 10, Rating: 3.3}},
-		{Title: "Title2", Author: "Author1", Meta: bookStats{TotalReads: 20, Rating: 3.6}},
-		{Title: "Title3", Author: "Author1", Meta: bookStats{TotalReads: 30, Rating: 3.9}},
-		{Title: "Title4", Author: "Author2", Meta: bookStats{TotalReads: 114, Rating: 4.0}},
-		{Title: "Title5", Author: "Author3", Meta: bookStats{TotalReads: 500, Rating: 4.8}},
+	jsonSampleData = []string{
+		`{"ID": "", "Title": "Title1", "Banned": true,  "Author": "Author1", "Meta": { "TotalReads": 100, "Rating": 3.2 }}`,
+		`{"ID": "", "Title": "Title2", "Banned": false, "Author": "Author1", "Meta": { "TotalReads": 150, "Rating": 4.1 }}`,
+		`{"ID": "", "Title": "Title3", "Banned": false, "Author": "Author2", "Meta": { "TotalReads": 120, "Rating": 4.6 }}`,
+		`{"ID": "", "Title": "Title4", "Banned": true,  "Author": "Author3", "Meta": { "TotalReads": 1000, "Rating": 2.6 }}`,
 	}
 
-	queries = []queryTest{
-		{name: "AllNil", query: nil, resIdx: []int{0, 1, 2, 3, 4}},
-		{name: "AllExplicit", query: &Query{}, resIdx: []int{0, 1, 2, 3, 4}},
+	boolTrue   = true
+	boolFalse  = false
+	title      = "Title"
+	title0     = "Title1"
+	title1     = "Title2"
+	title3     = "Title3"
+	titleMax   = "Title5000"
+	totreadEq1 = 100.0
+	totreadMid = 124.36
+	totreadEq2 = 150.0
+	totreadMax = 1221.25
+	totreadMin = 12.22
+	ratingMin  = 1.0
+	rating1    = 3.2
+	ratingMid  = 3.6
+	ratingMax  = 5.0
 
-		{name: "FromAuthor1", query: Where("Author").Eq("Author1"), resIdx: []int{0, 1, 2}},
-		{name: "FromAuthor2", query: Where("Author").Eq("Author2"), resIdx: []int{3}},
-		{name: "FromAuthor3", query: Where("Author").Eq("Author3"), resIdx: []int{4}},
+	jsonQueries = []jsonQueryTest{
+		// Ands string
+		{
+			name:   "EqByTitle1",
+			resIdx: []int{0},
+			query:  Where("Title").Eq(&title0),
+		},
+		{
+			name:   "NeByTitle1",
+			resIdx: []int{1, 2, 3},
+			query:  Where("Title").Ne(&title0),
+		},
+		{
+			name:   "NeByTitle",
+			resIdx: []int{0, 1, 2, 3},
+			query:  Where("Title").Ne(&title),
+		},
+		{
+			name:   "EqByTitle",
+			resIdx: []int{},
+			query:  Where("Title").Eq(&title),
+		},
+		{
+			name:   "LtByTitle",
+			resIdx: []int{},
+			query:  Where("Title").Lt(&title),
+		},
+		{
+			name:   "GtByTitle",
+			resIdx: []int{0, 1, 2, 3},
+			query:  Where("Title").Gt(&title),
+		},
+		{
+			name:   "GtByTitleMax",
+			resIdx: []int{},
+			query:  Where("Title").Gt(&titleMax),
+		},
 
-		{name: "AndAuthor1Title2", query: Where("Author").Eq("Author1").And("Title").Eq("Title2"), resIdx: []int{1}},
-		{name: "AndAuthorNestedTotalReads", query: Where("Author").Eq("Author1").And("Meta.TotalReads").Eq(float64(10)), resIdx: []int{0}},
+		// Ands "int" (which query interpret as float)
+		{
+			name:   "EqByTotalReads1",
+			resIdx: []int{0},
+			query:  Where("Meta.TotalReads").Eq(&totreadEq1),
+		},
+		{
+			name:   "LtByTotalReads1",
+			resIdx: []int{},
+			query:  Where("Meta.TotalReads").Lt(&totreadEq1),
+		},
+		{
+			name:   "LeByTotalReadsBigger",
+			resIdx: []int{0, 1, 2, 3},
+			query:  Where("Meta.TotalReads").Le(&totreadMax),
+		},
+		{
+			name:   "GeByTotalReadsMin",
+			resIdx: []int{0, 1, 2, 3},
+			query:  Where("Meta.TotalReads").Ge(&totreadMin),
+		},
+		{
+			name:   "LtByTotalReadsMidpoint",
+			resIdx: []int{0, 2},
+			query:  Where("Meta.TotalReads").Lt(&totreadMid),
+		},
+		{
+			name:   "GtByTotalReadsMidpoint",
+			resIdx: []int{1, 3},
+			query:  Where("Meta.TotalReads").Gt(&totreadMid),
+		},
+		{
+			name:   "LtByTotalReads2",
+			resIdx: []int{0, 2},
+			query:  Where("Meta.TotalReads").Lt(&totreadEq2),
+		},
 
-		{name: "OrAuthor", query: Where("Author").Eq("Author1").Or(Where("Author").Eq("Author3")), resIdx: []int{0, 1, 2, 4}},
+		// Ands float
+		{
+			name:   "EqByRating1",
+			resIdx: []int{0},
+			query:  Where("Meta.Rating").Eq(&rating1),
+		},
+		{
+			name:   "GtByRating1",
+			resIdx: []int{1, 2},
+			query:  Where("Meta.Rating").Gt(&rating1),
+		},
+		{
+			name:   "LeByRatingMin",
+			resIdx: []int{},
+			query:  Where("Meta.Rating").Le(&ratingMin),
+		},
+		{
+			name:   "GeByRatingMin",
+			resIdx: []int{0, 1, 2, 3},
+			query:  Where("Meta.Rating").Ge(&ratingMin),
+		},
+		{
+			name:   "LeByRatingMid",
+			resIdx: []int{0, 3},
+			query:  Where("Meta.Rating").Le(&ratingMid),
+		},
+		{
+			name:   "GeByRatingMid",
+			resIdx: []int{1, 2},
+			query:  Where("Meta.Rating").Ge(&ratingMid),
+		},
+		{
+			name:   "LeByRatingMax",
+			resIdx: []int{0, 1, 2, 3},
+			query:  Where("Meta.Rating").Le(&ratingMax),
+		},
+		{
+			name:   "GtByRatingMax",
+			resIdx: []int{},
+			query:  Where("Meta.Rating").Gt(&ratingMax),
+		},
+		// Ands bool
+		{
+			name:   "EqByBanned",
+			resIdx: []int{0, 3},
+			query:  Where("Banned").Eq(&boolTrue),
+		},
+		{
+			name:   "NeByBanned",
+			resIdx: []int{1, 2},
+			query:  Where("Banned").Ne(&boolTrue),
+		},
+		{
+			name:   "EqByBannedFalse",
+			resIdx: []int{1, 2},
+			query:  Where("Banned").Eq(&boolFalse),
+		},
 
-		{name: "NeAuthor", query: Where("Author").Ne("Author1"), resIdx: []int{3, 4}},
-		{name: "NeTotalReads", query: Where("Meta.TotalReads").Ne(float64(30)), resIdx: []int{0, 1, 3, 4}},
-		{name: "NeRating", query: Where("Meta.Rating").Ne(3.6), resIdx: []int{0, 2, 3, 4}},
+		// Ors
+		{
+			name:   "EqTitle1OrTitle3",
+			resIdx: []int{0, 2},
+			query:  Where("Title").Eq(&title0).Or(Where("Title").Eq(&title3)),
+		},
+		{
+			name:   "EqTitle2OrRating",
+			resIdx: []int{0, 1},
+			query:  Where("Title").Eq(&title1).Or(Where("Meta.Rating").Eq(&rating1)),
+		},
 
-		{name: "GtAuthor", query: Where("Author").Gt("Author2"), resIdx: []int{4}},
-		{name: "GtTotalReads", query: Where("Meta.TotalReads").Gt(float64(30)), resIdx: []int{3, 4}},
-		{name: "GtRating", query: Where("Meta.Rating").Gt(3.6), resIdx: []int{2, 3, 4}},
-
-		{name: "GeAuthor", query: Where("Author").Ge("Author2"), resIdx: []int{3, 4}},
-		{name: "GeTotalReads", query: Where("Meta.TotalReads").Ge(float64(30)), resIdx: []int{2, 3, 4}},
-		{name: "GeRating", query: Where("Meta.Rating").Ge(3.6), resIdx: []int{1, 2, 3, 4}},
-
-		{name: "LtAuthor", query: Where("Author").Lt("Author2"), resIdx: []int{0, 1, 2}},
-		{name: "LtTotalReads", query: Where("Meta.TotalReads").Lt(float64(30)), resIdx: []int{0, 1}},
-		{name: "LtRating", query: Where("Meta.Rating").Lt(3.6), resIdx: []int{0}},
-
-		{name: "LeAuthor", query: Where("Author").Le("Author2"), resIdx: []int{0, 1, 2, 3}},
-		{name: "LeTotalReads", query: Where("Meta.TotalReads").Le(float64(30)), resIdx: []int{0, 1, 2}},
-		{name: "LeRating", query: Where("Meta.Rating").Le(3.6), resIdx: []int{0, 1}},
-
-		{name: "SortAscString", query: Where("Meta.TotalReads").Gt(float64(20)).OrderBy("Author"), resIdx: []int{2, 3, 4}, ordered: true},
-		{name: "SortDescString", query: Where("Meta.TotalReads").Gt(float64(20)).OrderByDesc("Author"), resIdx: []int{4, 3, 2}, ordered: true},
-
-		{name: "SortAscInt", query: Where("Meta.TotalReads").Gt(float64(10)).OrderBy("Meta.TotalReads"), resIdx: []int{1, 2, 3, 4}, ordered: true},
-		{name: "SortDescInt", query: Where("Meta.TotalReads").Gt(float64(10)).OrderByDesc("Meta.TotalReads"), resIdx: []int{4, 3, 2, 1}, ordered: true},
-
-		{name: "SortAscFloat", query: Where("Meta.TotalReads").Gt(float64(10)).OrderBy("Meta.Rating"), resIdx: []int{1, 2, 3, 4}, ordered: true},
-		{name: "SortDescFloat", query: Where("Meta.TotalReads").Gt(float64(10)).OrderByDesc("Meta.Rating"), resIdx: []int{4, 3, 2, 1}, ordered: true},
-
-		{name: "SortAllAscString", query: OrderBy("Title"), resIdx: []int{0, 1, 2, 3, 4}, ordered: true},
-		{name: "SortAllDescString", query: OrderByDesc("Title"), resIdx: []int{4, 3, 2, 1, 0}, ordered: true},
-
-		{name: "SortAllAscInt", query: OrderBy("Meta.TotalReads"), resIdx: []int{0, 1, 2, 3, 4}, ordered: true},
-		{name: "SortAllDescInt", query: OrderByDesc("Meta.TotalReads"), resIdx: []int{4, 3, 2, 1, 0}, ordered: true},
-
-		{name: "SortAllAscFloat", query: OrderBy("Meta.Rating"), resIdx: []int{0, 1, 2, 3, 4}, ordered: true},
-		{name: "SortAllDescFloat", query: OrderByDesc("Meta.Rating"), resIdx: []int{4, 3, 2, 1, 0}, ordered: true},
+		// Ordering (string, int, float)
+		{
+			name:    "AllOrderedTitle",
+			resIdx:  []int{0, 1, 2, 3},
+			query:   OrderBy("Title"),
+			ordered: true,
+		},
+		{
+			name:    "AllOrderedTitleDesc",
+			resIdx:  []int{3, 2, 1, 0},
+			query:   OrderByDesc("Title"),
+			ordered: true,
+		},
+		{
+			name:    "AllOrderedTotalReads",
+			resIdx:  []int{0, 2, 1, 3},
+			query:   OrderBy("Meta.TotalReads"),
+			ordered: true,
+		},
+		{
+			name:    "AllOrderedTotalReads",
+			resIdx:  []int{3, 1, 2, 0},
+			query:   OrderByDesc("Meta.TotalReads"),
+			ordered: true,
+		},
+		{
+			name:    "AllOrderedRatings",
+			resIdx:  []int{3, 0, 1, 2},
+			query:   OrderBy("Meta.Rating"),
+			ordered: true,
+		},
+		{
+			name:    "AllOrderedRatingsDesc",
+			resIdx:  []int{2, 1, 0, 3},
+			query:   OrderByDesc("Meta.Rating"),
+			ordered: true,
+		},
+		{
+			name:    "AllOrderedRatingsDescWithAnd",
+			resIdx:  []int{2, 1},
+			query:   Where("Meta.Rating").Gt(&ratingMid).OrderBy("Meta.TotalReads"),
+			ordered: true,
+		},
+		{
+			name:    "AllOrderedRatingsDescWithAndDesc",
+			resIdx:  []int{1, 2},
+			query:   Where("Meta.Rating").Gt(&ratingMid).OrderByDesc("Meta.TotalReads"),
+			ordered: true,
+		},
+		// Indexing
+		{
+			name:   "EqTitle1OrTitle3UseIndex",
+			resIdx: []int{0, 2},
+			query:  Where("Title").Eq(&title0).Or(Where("Title").Eq(&title3)).UseIndex("Title"),
+		},
+		{
+			name:   "GeByTotalReadsMinUseIndex",
+			resIdx: []int{0, 1, 2, 3},
+			query:  Where("Meta.TotalReads").Ge(&totreadMin).UseIndex("Meta.TotalReads"),
+		},
+		{
+			name:   "InvalidIndex",
+			resIdx: []int{},
+			query:  Where("Meta.TotalReads").Ge(&totreadMin).UseIndex("Not.Valid.Path"),
+		},
 	}
 )
 
-func TestCollectionQuery(t *testing.T) {
-	t.Parallel()
-	c, clean := createCollectionWithData(t)
+func TestQuery(t *testing.T) {
+	c, clean := createCollectionWithJSONData(t)
 	defer clean()
-	for _, q := range queries {
+
+	for _, q := range jsonQueries {
 		q := q
 		t.Run(q.name, func(t *testing.T) {
 			t.Parallel()
-			ret, err := c.Find(q.query)
-			if err != nil {
-				t.Fatalf("error when executing query: %v", err)
+			res, err := c.Find(q.query)
+			checkErr(t, err)
+			if len(q.resIdx) != len(res) {
+				t.Fatalf("query results length doesn't match, expected: %d, got: %d", len(q.resIdx), len(res))
 			}
-			if len(q.resIdx) != len(ret) {
-				t.Fatalf("query results length doesn't match, expected: %d, got: %d", len(q.resIdx), len(ret))
-			}
-			res := make([]*book, len(ret))
-			for i, bookJSON := range ret {
-				var book = &book{}
-				util.InstanceFromJSONString(bookJSON, book)
-				res[i] = book
-			}
-
 			expectedIdx := make([]int, len(q.resIdx))
 			for i := range q.resIdx {
 				expectedIdx[i] = q.resIdx[i]
 			}
 			if !q.ordered {
 				sort.Slice(res, func(i, j int) bool {
-					return strings.Compare(res[i].ID.String(), res[j].ID.String()) == -1
+					return strings.Compare(res[i], res[j]) == -1
 				})
 				sort.Slice(expectedIdx, func(i, j int) bool {
-					return strings.Compare(sampleData[expectedIdx[i]].ID.String(), sampleData[expectedIdx[j]].ID.String()) == -1
+					return strings.Compare(jsonSampleData[expectedIdx[i]], jsonSampleData[expectedIdx[j]]) == -1
 				})
 			}
 			for i, idx := range expectedIdx {
-				if !reflect.DeepEqual(sampleData[idx], *res[i]) {
-					t.Fatalf("wrong query item result, expected: %v, got: %v", sampleData[idx], *res[i])
+				if jsonSampleData[idx] != res[i] {
+					t.Fatalf("wrong query item result, expected: %v, got: %v", jsonSampleData[idx], res[i])
 				}
 			}
 		})
 	}
 }
 
-func TestInvalidSortField(t *testing.T) {
-	t.Parallel()
-
-	c, clean := createCollectionWithData(t)
-	defer clean()
-	_, err := c.Find(OrderBy("WrongFieldName"))
-	if !errors.Is(err, ErrInvalidSortingField) {
-		t.Fatal("query should fail using an invalid field")
-	}
-}
-
-func createCollectionWithData(t *testing.T) (*Collection, func()) {
-	db, clean := createTestDB(t)
-	c, err := db.NewCollection(CollectionConfig{
+func createCollectionWithJSONData(t *testing.T) (*Collection, func()) {
+	s, clean := createTestDB(t)
+	c, err := s.NewCollection(CollectionConfig{
 		Name:   "Book",
-		Schema: util.SchemaFromInstance(&book{}),
+		Schema: testQuerySchema,
+		Indexes: []IndexConfig{{
+			Path: "Meta.TotalReads",
+		},
+			{
+				Path: "Title",
+			}},
 	})
 	checkErr(t, err)
-	for i := range sampleData {
-		sampleDataString := util.JSONStringFromInstance(&sampleData[i])
-		if err = c.Create(sampleDataString); err != nil {
+	for i := range jsonSampleData {
+		if err = c.Create(&jsonSampleData[i]); err != nil {
 			t.Fatalf("failed to create sample data: %v", err)
 		}
-		updated := book{}
-		util.InstanceFromJSONString(*sampleDataString, &updated)
-		sampleData[i] = updated
 	}
 	return c, clean
 }
